+++ QuizApplication/app/src/main/java/ro/ase/codinquiz/quizapplication/Main/Teacher/Adapters/FinishedTestListAdapter.java
@@ -32,7 +32,7 @@ public class FinishedTestListAdapter extends ArrayAdapter<FinishedTest> {
     public View getView(int position, @Nullable View convertView, @NonNull ViewGroup parent) {
         String username=getItem(position).getUsername();
         String testName=getItem(position).getTestName();
-        float score=getItem(position).getScore();
+        int score=getItem(position).getScore();
         Calendar calendar= new GregorianCalendar(2019,1,1);
      FinishedTest finishedTest=new FinishedTest(1,1,score,username,testName,calendar.getTime());
         LayoutInflater inflater=LayoutInflater.from(context);
@@ -43,7 +43,7 @@ public class FinishedTestListAdapter extends ArrayAdapter<FinishedTest> {
 
         textViewTest.setText(testName);
         textViewName.setText(username);
-        String s = Float.toString(score);
+        String s = Integer.toString(score);
         textViewGrade.setText(s);
 
         return convertView;
